introduce separate ItemClasses file
